Make sure we can pass a list of files into the replacer.
While we're at it, verify the inputs are actually files, and benchmark
performance - CurationTasks relies on things not being too slow.

+++ lib/replacer.rb
@@ -14,36 +14,36 @@ class Replacer
 
   def update(files)
     files.each do |filename|
-      @files_processed[:filename] = false
-      begin
-        temp_file = Tempfile.new('closed_courseware')
+      @files_processed[filename] = false
+      if File.file?(filename)
+        begin
+          temp_file = Tempfile.new('closed_courseware')
 
-        File.open(filename, 'r+') do |file|
-          file.each_line do |line|
-            # Execution of the code block will implicitly return the last line,
-            # which is then used as the replacement pattern. The processed
-            # variable must be incremented *first* or it will be returned and
-            # we will change URLs into numbers.
-            subbed_line = line.gsub!('http://ocw.mit.edu') {
-              @links_processed += 1; 'https://ocw.mit.edu';
-            }
+          File.open(filename, 'r+') do |file|
+            file.each_line do |line|
+              # Execution of the code block will implicitly return the last line,
+              # which is then used as the replacement pattern. The processed
+              # variable must be incremented *first* or it will be returned and
+              # we will change URLs into numbers.
+              subbed_line = line.gsub!('http://ocw.mit.edu') {
+                @links_processed += 1; 'https://ocw.mit.edu';
+              }
 
-            # gsub returns nil if it doesn't make any substitutions. In that
-            # case, we want to use the original line.
-            temp_file.puts subbed_line ? subbed_line : line
+              # gsub returns nil if it doesn't make any substitutions. In that
+              # case, we want to use the original line.
+              temp_file.puts subbed_line ? subbed_line : line
+            end
           end
-        end
 
-        temp_file.close
-        FileUtils.mv(temp_file.path, filename)
-        @files_processed[:filename] = true
-      ensure
-        temp_file.close
-        temp_file.unlink
+          temp_file.close
+          FileUtils.mv(temp_file.path, filename)
+          @files_processed[filename] = true
+        ensure
+          temp_file.close
+          temp_file.unlink
+        end
       end
     end
-
-    puts @links_processed.to_s + ' link' + (@links_processed == 1 ? '' : 's') + ' processed. You so fancy!'
   end
 
 end
